BulletinBoard/requirements.txt
APScheduler==3.10.1
asgiref==3.6.0
bleach==6.0.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
defusedxml==0.7.1
Django==4.2.1
django-allauth==0.54.0
django-apscheduler==0.6.2
django-filter==23.2
django-summernote==0.8.20.0
idna==3.4
oauthlib==3.2.2
pycparser==2.21
PyJWT==2.6.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
requests==2.30.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
tzlocal==4.3
urllib3==2.0.2
webencodings==0.5.1
